index_scan: merge the spec-side bridge back into spec.c
The previous commit put the spec-side scan glue in a separate spec_scan.c, but
that created a spec <-> spec_scan symbol cycle: spec.c triggers/queries scans
(FT.CREATE/ALTER, RDB load, INFO) while the bridge calls back into spec to index.
The cycle was the tell that spec and the bridge are really one layer.

Fold the bridge (per-key callbacks, IndexSpec_ScanAndReindex / Indexes_ScanAndReindex,
the scanner<->spec linkage and the global scan-all singleton, the FT.INFO helpers)
back into spec.c and delete spec_scan.{c,h}. The valuable split is preserved: the
generic, reusable engine stays in index_scan.{c,h}, still free of any spec dependency
(`nm -u index_scan.c.o` shows zero spec symbols), and remains where the upcoming
async-scan driver will live.

Resulting dependency graph is acyclic: spec.c -> index_scan (one-way), index_scan a
leaf. No behavior change.
